separate fetch and filter relics into hooks
